+++ src/main/java/io/kaitai/struct/RandomAccessFileKaitaiStream.java
@@ -338,7 +338,7 @@ public class RandomAccessFileKaitaiStream extends KaitaiStream {
     }
 
     @Override
-    public byte[] readBytesTerm(int term, boolean includeTerm, boolean consumeTerm, boolean eosError) {
+    public byte[] readBytesTerm(byte term, boolean includeTerm, boolean consumeTerm, boolean eosError) {
         try {
             ByteArrayOutputStream buf = new ByteArrayOutputStream();
             while (true) {
@@ -349,7 +349,7 @@ public class RandomAccessFileKaitaiStream extends KaitaiStream {
                     } else {
                         return buf.toByteArray();
                     }
-                } else if (c == term) {
+                } else if ((byte) c == term) {
                     if (includeTerm)
                         buf.write(c);
                     if (!consumeTerm)
